requirements.txt
Django==5.0.6
pillow==10.4.0
django-notifications-hq==1.8.3
setuptools==70.2.0